Remove IBTK::FEMapCache.
We don't need this now that we have FEMapping et al.
