+++ doc/news/changes/incompatibilities/20200901DavidWells
@@ -1,0 +1,3 @@
+Removed: The class IBTK::FEMapCache has been removed.
+<br>
+(David Wells, 2020/09/01)
