+++ tests/IBTK/ghost_accumulation_01.cpp
@@ -73,8 +73,10 @@ main(int argc, char* argv[])
         const IntVector<NDIM> gcw(input_db->getIntegerWithDefault("ghost_width", 1));
         plog << "GCW: " << gcw << '\n';
         const bool use_cell = input_db->getStringWithDefault("var_type", "CELL") == "CELL";
-        auto u_var = use_cell ? Pointer<Variable<NDIM> >(new CellVariable<NDIM, double>("u", u_depth)) :
-                                Pointer<Variable<NDIM> >(new SideVariable<NDIM, double>("u", u_depth));
+
+        // disambiguate between libMesh::Variable and SAMRAI::hier::Variable
+        auto u_var = use_cell ? Pointer<hier::Variable<NDIM> >(new CellVariable<NDIM, double>("u", u_depth)) :
+                                Pointer<hier::Variable<NDIM> >(new SideVariable<NDIM, double>("u", u_depth));
 
         const int u_idx = var_db->registerVariableAndContext(u_var, ctx, gcw);
